# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Untukmu API",
    "description": "Untukmu Fase 1–5: akun, pesan terenkripsi, sosial, doa, kurasi, pemulihan dan ekspor.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Akun dan sesi"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Akun dan sesi"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/sessions": {
      "get": {
        "tags": [
          "Akun dan sesi"
        ],
        "summary": "Sessions",
        "operationId": "sessions_api_v1_auth_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/sessions/{session_id}": {
      "delete": {
        "tags": [
          "Akun dan sesi"
        ],
        "summary": "Revoke",
        "operationId": "revoke_api_v1_auth_sessions__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Akun dan sesi"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/settings": {
      "patch": {
        "tags": [
          "Akun dan sesi"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_v1_users_me_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/recovery/enable": {
      "post": {
        "tags": [
          "Akun dan sesi"
        ],
        "summary": "Enable Recovery",
        "operationId": "enable_recovery_api_v1_auth_recovery_enable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableRecovery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/clear": {
      "post": {
        "tags": [
          "Akun dan sesi"
        ],
        "summary": "Clear Data",
        "operationId": "clear_data_api_v1_users_me_clear_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmCredential"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmCredential": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "password"
        ],
        "title": "ConfirmCredential"
      },
...
      "EnableRecovery": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          },
          "recovery_verifier": {
            "type": "string",
            "pattern": "^[A-Za-z0-9+/]{43}=$",
            "title": "Recovery Verifier"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "password",
          "recovery_verifier"
        ],
        "title": "EnableRecovery"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KeyRecord": {
        "properties": {
          "v": {
            "type": "integer",
            "const": 2,
            "title": "V",
            "default": 2
          },
          "kdf": {
            "type": "string",
            "const": "Argon2id",
            "title": "Kdf",
            "default": "Argon2id"
          },
          "memory": {
            "type": "integer",
            "const": 65536,
            "title": "Memory",
            "default": 65536
          },
          "iter": {
            "type": "integer",
            "const": 3,
            "title": "Iter",
            "default": 3
          },
          "parallelism": {
            "type": "integer",
            "const": 1,
            "title": "Parallelism",
            "default": 1
          },
          "salt": {
            "type": "string",
            "title": "Salt"
          },
          "wrapped": {
            "$ref": "#/components/schemas/Wrapped"
          },
          "wrappedRecovery": {
            "$ref": "#/components/schemas/Wrapped"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "salt",
          "wrapped",
          "wrappedRecovery"
        ],
        "title": "KeyRecord"
      },
      "Login": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Login"
      },
...
      "Register": {
        "properties": {
          "recovery_verifier": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9+/]{43}=$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Verifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 12,
            "title": "Password"
          },
          "display_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "default": "Anonim"
          },
          "encryption_record": {
            "$ref": "#/components/schemas/KeyRecord"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password",
          "encryption_record"
        ],
        "title": "Register"
      },
...
      "UserSettings": {
        "properties": {
          "default_message_visibility": {
            "type": "string",
            "enum": [
              "private",
              "public_anon",
              "unlisted"
            ],
            "title": "Default Message Visibility",
            "default": "private"
          },
          "profile_visibility": {
            "type": "string",
            "enum": [
              "private",
              "public"
            ],
            "title": "Profile Visibility",
            "default": "private"
          },
          "galaxy_background": {
            "type": "boolean",
            "title": "Galaxy Background",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UserSettings"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Wrapped": {
        "properties": {
          "iv": {
            "type": "string",
            "title": "Iv"
          },
          "ct": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Ct"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "iv",
          "ct"
        ],
        "title": "Wrapped"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}